Each blended feature now has it's own project and GAV
