+++ build.sbt
@@ -17,7 +17,43 @@ lazy val global = Def.settings(
   Global / resolvers += "SpringExternal" at "http://repository.springsource.com/maven/bundles/external"
 )
 
-lazy val blendedLauncherFeatures = BlendedLauncherFeatures.project
+lazy val blendedLauncherFeatureActivemq = BlendedLauncherFeatureActivemq.project
+lazy val blendedLauncherFeatureAkkaHttp = BlendedLauncherFeatureAkkaHttp.project
+lazy val blendedLauncherFeatureBase = BlendedLauncherFeatureBase.project
+lazy val blendedLauncherFeatureBaseEquinox = BlendedLauncherFeatureBaseEquinox.project
+lazy val blendedLauncherFeatureBaseFelix = BlendedLauncherFeatureBaseFelix.project
+lazy val blendedLauncherFeatureCamel = BlendedLauncherFeatureCamel.project
+lazy val blendedLauncherFeatureCommons= BlendedLauncherFeatureCommons.project
+lazy val blendedLauncherFeatureHawtio = BlendedLauncherFeatureHawtio.project
+lazy val blendedLauncherFeatureJetty = BlendedLauncherFeatureJetty.project
+lazy val blendedLauncherFeatureMgmtClient = BlendedLauncherFeatureMgmtClient.project
+lazy val blendedLauncherFeatureMgmtServer = BlendedLauncherFeatureMgmtServer.project
+lazy val blendedLauncherFeaturePersistence = BlendedLauncherFeaturePersistence.project
+lazy val blendedLauncherFeatureSamples = BlendedLauncherFeatureSamples.project
+lazy val blendedLauncherFeatureSecurity = BlendedLauncherFeatureSecurity.project
+lazy val blendedLauncherFeatureSpring = BlendedLauncherFeatureSpring.project
+lazy val blendedLauncherFeatureSsl = BlendedLauncherFeatureSsl.project
+lazy val blendedLauncherFeatureStreams = BlendedLauncherFeatureStreams.project
+
+lazy val blendedLauncherFeature = project.settings(CommonSettings()).aggregate(
+  blendedLauncherFeatureActivemq,
+  blendedLauncherFeatureAkkaHttp,
+  blendedLauncherFeatureBase,
+  blendedLauncherFeatureBaseEquinox,
+  blendedLauncherFeatureBaseFelix,
+  blendedLauncherFeatureCamel,
+  blendedLauncherFeatureHawtio,
+  blendedLauncherFeatureJetty,
+  blendedLauncherFeatureMgmtClient,
+  blendedLauncherFeatureMgmtServer,
+  blendedLauncherFeaturePersistence,
+  blendedLauncherFeatureSamples,
+  blendedLauncherFeatureSecurity,
+  blendedLauncherFeatureSpring,
+  blendedLauncherFeatureSsl,
+  blendedLauncherFeatureStreams
+)
+
 lazy val blendedDemoNodeResources = BlendedDemoNodeResources.project
 lazy val blendedDemoMgmtResources = BlendedDemoMgmtResources.project
 lazy val blendedDemoNode = BlendedDemoNode.project
@@ -34,7 +70,7 @@ lazy val root = {
     .settings(global)
     .settings(PublishConfig.doPublish)
     .aggregate(
-      blendedLauncherFeatures,
+      blendedLauncherFeature,
       blendedDemoNodeResources,
       blendedDemoMgmtResources,
       blendedDemoNode,
