+++ container/blended.demo.mgmt/src/main/resources/profile/profile.conf
@@ -72,18 +72,18 @@ resources = [
 ]
 
 features = [
-  {name = blended-base-felix, version = "${blended.version}"},
-  // {name = blended-base-equinox, version = "${blended.version}"},
-  {name = blended-base, version = "${blended.version}"},
-  {name = blended-commons, version = "${blended.version}"},
-  {name = blended-jetty, version = "${blended.version}"},
-  {name = blended-hawtio, version = "${blended.version}"},
-  {name = blended-spring, version = "${blended.version}"},
-  {name = blended-security, version = "${blended.version}"},
-  {name = blended-mgmt-client, version = "${blended.version}"},
-  {name = blended-mgmt-server, version = "${blended.version}"},
-  {name = blended-akka-http, version = "${blended.version}"},
-  {name = blended-persistence, version = "${blended.version}"}
+  {name = blended.launcher.feature.base.felix, version = "${blended.version}"},
+  // {name = blended.launcher.feature.base.equinox, version = "${blended.version}"},
+  {name = blended.launcher.feature.base, version = "${blended.version}"},
+  {name = blended.launcher.feature.commons, version = "${blended.version}"},
+  {name = blended.launcher.feature.jetty, version = "${blended.version}"},
+  {name = blended.launcher.feature.hawtio, version = "${blended.version}"},
+  {name = blended.launcher.feature.spring, version = "${blended.version}"},
+  {name = blended.launcher.feature.security, version = "${blended.version}"},
+  {name = blended.launcher.feature.mgmt.client, version = "${blended.version}"},
+  {name = blended.launcher.feature.mgmt.server, version = "${blended.version}"},
+  {name = blended.launcher.feature.akka.http, version = "${blended.version}"},
+  {name = blended.launcher.feature.persistence, version = "${blended.version}"}
 ]
 
 bundles = []
