+++ container/blended.demo.node/src/main/resources/profile/profile.conf
@@ -66,22 +66,22 @@ resources=[
 ]
 
 features=[
-    { name=blended-base-felix, version="${blended.version}"}
-    // { name=blended-base-equinox, version="${blended.version}"},
-    { name=blended-base, version="${blended.version}"},
-    { name=blended-commons, version="${blended.version}" },
-    { name=blended-spring, version="${blended.version}" },
-    { name=blended-ssl, version="${blended.version}" },
-    { name=blended-jetty, version="${blended.version}" },
-    { name=blended-hawtio, version="${blended.version}" },
-    { name=blended-activemq, version="${blended.version}" },
-    { name=blended-camel, version = "${blended.version}" },
-    { name=blended-security, version = "${blended.version}" },
-    { name=blended-mgmt-client, version = "${blended.version}" },
-    { name=blended-akka-http, version = "${blended.version}" },
-    { name=blended-persistence, version="${blended.version}" },
-    { name=blended-streams, version="${blended.version}" },
-    { name=blended-samples, version="${blended.version}" }
+    { name=blended.launcher.feature.base.felix, version="${blended.version}"}
+    // { name=blended.launcher.feature.base.equinox, version="${blended.version}"},
+    { name=blended.launcher.feature.base, version="${blended.version}"},
+    { name=blended.launcher.feature.commons, version="${blended.version}" },
+    { name=blended.launcher.feature.spring, version="${blended.version}" },
+    { name=blended.launcher.feature.ssl, version="${blended.version}" },
+    { name=blended.launcher.feature.jetty, version="${blended.version}" },
+    { name=blended.launcher.feature.hawtio, version="${blended.version}" },
+    { name=blended.launcher.feature.activemq, version="${blended.version}" },
+    { name=blended.launcher.feature.camel, version = "${blended.version}" },
+    { name=blended.launcher.feature.security, version = "${blended.version}" },
+    { name=blended.launcher.feature.mgmt.client, version = "${blended.version}" },
+    { name=blended.launcher.feature.akka.http, version = "${blended.version}" },
+    { name=blended.launcher.feature.persistence, version="${blended.version}" },
+    { name=blended.launcher.feature.streams, version="${blended.version}" },
+    { name=blended.launcher.feature.samples, version="${blended.version}" }
 ]
 bundles=[
 ]
